007_stock_trade/requirements.txt
PyQt5
pandas
python-dotenv
